fix: tabela de edicao_sobre_mims deletada (não tinha visto que já existia a tabela sobre_mims)
