--- app/DTO/EdicaoSobreMimDTO.ts
@@ -1,9 +1,0 @@
-export default interface EdicaoSobreMimDTO {
-  id?: string
-  titulo_um?: string
-  id_imagem_um?: string
-  texto_um?: string
-  titulo_dois?: string
-  id_imagem_dois?: string
-  texto_dois?: string
-}
