+++ app/Repositories/SobreMimsRepository.ts
@@ -1,8 +1,7 @@
 import Database from '@ioc:Adonis/Lucid/Database'
-import SobreMimsDTO from 'App/DTO/SobreMimsDTO'
-
-export default class SobreMimsRepository {
-  public static async find(params: SobreMimsDTO) {
+import SobreMimDTO from 'App/DTO/SobreMimsDTO'
+export default class SobreMimRepository {
+  public static async find(params: SobreMimDTO): Promise<SobreMimDTO[]> {
     const result = await Database.query().from('sobre_mims').where(params)
 
     return result
