+++ start/routes.ts
@@ -59,10 +59,10 @@ Route.post('/comentarios', 'ComentariosController.store')
 Route.put('/comentarios/:id', 'ComentariosController.update')
 Route.delete('/comentarios/:id', 'ComentariosController.destroy')
 
-Route.get('/sobremims', 'SobreMimsController.index')
-Route.post('/sobremims', 'SobreMimsController.store')
-Route.put('/sobremims/:id', 'SobreMimsController.update')
-Route.delete('/sobremims/:id', 'SobreMimsController.destroy')
+Route.get('/sobre_mims', 'SobreMimsController.index')
+Route.post('/sobre_mims', 'SobreMimsController.store')
+Route.put('/sobre_mims/:id', 'SobreMimsController.update')
+Route.delete('/sobre_mims/:id', 'SobreMimsController.destroy')
 
 Route.get('/consultorios', 'ConsultoriosController.index')
 Route.get('/consultorios/:id', 'ConsultoriosController.indexByIdConsultorio')
@@ -166,7 +166,3 @@ Route.post('/medicos_indicados', 'MedicosIndicadosController.store')
 Route.put('/medicos_indicados/:id', 'MedicosIndicadosController.update')
 Route.delete('/medicos_indicados/:id', 'MedicosIndicadosController.destroy')
 
-Route.get('/edicao_sobre_mims', 'EdicaoSobreMimsController.index')
-Route.post('/edicao_sobre_mims', 'EdicaoSobreMimsController.store')
-Route.put('/edicao_sobre_mims/:id', 'EdicaoSobreMimsController.update')
-Route.delete('/edicao_sobre_mims/:id', 'EdicaoSobreMimsController.destroy')
